requirements.txt
PyAutoGUI==0.9.54
pyttsx3==2.90
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Accessibility==10.1
Pillow==10.2.0
psutil==5.9.8
